# internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proton Mail Tools",
    "version": "0.1.0",
    "description": "Tools for reading, searching, sending and organising email in a Proton Mail account via Proton Mail Bridge. Messages are identified by a mailbox name plus a numeric uid returned by search_messages."
  },
  "security": [{"bearerAuth": []}],
  "paths": {
    "/mailboxes": {
      "get": {
        "operationId": "list_mailboxes",
        "summary": "List mailboxes",
        "description": "List every mailbox in the account with its kind (system, folder, or label), role (inbox, sent, trash, archive, spam, drafts, starred, all) and message counts. Folders hold a message exclusively (use move_messages); labels are additive tags (use label_messages). Use the exact 'name' value as the 'mailbox' argument for other tools.",
        "tags": ["mail"],
        "responses": {
          "200": {
            "description": "Mailboxes",
            "content": {"application/json": {"schema": {
              "type": "object",
              "properties": {"mailboxes": {"type": "array", "items": {"$ref": "#/components/schemas/Mailbox"}}}
            }}}
          }
        }
      }
    },
    "/messages/search": {
      "post": {
        "operationId": "search_messages",
        "summary": "Search messages",
        "description": "Search one mailbox (default INBOX) and return the newest matching messages first. All filters are optional and combine with AND. Returns summaries only; call read_message with the mailbox and uid to get the body.",
        "tags": ["mail"],
        "requestBody": {"required": false, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SearchRequest"}}}},
        "responses": {
          "200": {
            "description": "Matching messages, newest first",
            "content": {"application/json": {"schema": {
              "type": "object",
              "properties": {
                "count": {"type": "integer"},
                "messages": {"type": "array", "items": {"$ref": "#/components/schemas/MessageSummary"}}
              }
            }}}
          }
        }
      }
    },
    "/messages/read": {
      "post": {
        "operationId": "read_message",
        "summary": "Read a message",
        "description": "Fetch the full content of one message: headers, plain-text body (HTML is converted to text), the hyperlinks found in the HTML body, List-Unsubscribe URLs, and attachment metadata. Reading does not mark the message as read; use update_message_flags for that.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessageRef"}}}},
        "responses": {
          "200": {"description": "The message", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}},
          "404": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/messages/send": {
      "post": {
        "operationId": "send_message",
        "summary": "Send a new email",
        "description": "Compose and send a new plain-text email from the account owner. Confirm recipients, subject and body with the user before calling this.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SendRequest"}}}},
        "responses": {
          "200": {"description": "Sent", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SentResponse"}}}},
          "403": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/messages/reply": {
      "post": {
        "operationId": "reply_to_message",
        "summary": "Reply to a message",
        "description": "Send a reply to an existing message. Recipients, subject and threading headers are derived from the original automatically; supply only the reply body. Set reply_all to include everyone on the original thread.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ReplyRequest"}}}},
        "responses": {
          "200": {"description": "Sent", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SentResponse"}}}},
          "403": {"$ref": "#/components/responses/Error"},
          "404": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/messages/flags": {
      "post": {
        "operationId": "update_message_flags",
        "summary": "Mark messages read/unread or starred/unstarred",
        "description": "Set the read and/or starred state of one or more messages in a mailbox. 'flagged' is Proton's star: flagged=true stars a message (it appears under Starred), flagged=false unstars it. Omit a field to leave it unchanged.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FlagsRequest"}}}},
        "responses": {
          "200": {"description": "Updated", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusResponse"}}}}
        }
      }
    },
    "/messages/move": {
      "post": {
        "operationId": "move_messages",
        "summary": "Move messages to another mailbox",
        "description": "Move one or more messages into an existing mailbox, e.g. 'Archive' or 'Folders/Receipts'. Use list_mailboxes to find valid destination names.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MoveRequest"}}}},
        "responses": {
          "200": {"description": "Moved", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusResponse"}}}},
          "404": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/messages/archive": {
      "post": {
        "operationId": "archive_messages",
        "summary": "Archive messages",
        "description": "Move one or more messages out of their current mailbox into Archive. Messages already in Archive are reported in skipped_uids.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessagesRef"}}}},
        "responses": {
          "200": {"description": "Archived", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusResponse"}}}}
        }
      }
    },
    "/messages/label": {
      "post": {
        "operationId": "label_messages",
        "summary": "Apply a label (tag) to messages",
        "description": "Add a Proton label to one or more messages. Labels are additive tags: the message stays where it is and also appears under 'Labels/<name>'. Use list_mailboxes to see existing labels (kind = label); the label may be given as 'Receipts' or 'Labels/Receipts'. Messages that already have the label are reported in skipped_uids.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LabelRequest"}}}},
        "responses": {
          "200": {"description": "Labeled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LabelResponse"}}}},
          "404": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/messages/unlabel": {
      "post": {
        "operationId": "unlabel_messages",
        "summary": "Remove a label (tag) from messages",
        "description": "Remove a Proton label from one or more messages, identified by their mailbox and uid (any mailbox they appear in). The messages themselves are not moved or deleted. Messages that do not carry the label are reported in skipped_uids.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LabelRequest"}}}},
        "responses": {
          "200": {"description": "Unlabeled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LabelResponse"}}}},
          "404": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/messages/delete": {
      "post": {
        "operationId": "delete_messages",
        "summary": "Permanently delete messages (Spam and Trash only)",
        "description": "Permanently and irreversibly delete one or more messages. Only allowed for messages in the Spam or Trash mailbox, matching Proton's own behaviour; for any other mailbox use trash_messages instead. Confirm with the user before calling this. Returns 403 if permanent deletion is disabled by configuration.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessagesRef"}}}},
        "responses": {
          "200": {"description": "Deleted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeleteResponse"}}}},
          "400": {"$ref": "#/components/responses/Error"},
          "403": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/mailboxes/empty": {
      "post": {
        "operationId": "empty_mailbox",
        "summary": "Permanently delete every message in Spam or Trash",
        "description": "Empty the Spam or Trash mailbox, permanently deleting all messages in it. Not allowed for any other mailbox. Confirm with the user before calling this. Returns 403 if permanent deletion is disabled by configuration.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {
          "type": "object",
          "properties": {"mailbox": {"type": "string", "description": "'Spam' or 'Trash' (exact name from list_mailboxes)"}},
          "required": ["mailbox"]
        }}}},
        "responses": {
          "200": {"description": "Emptied", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeleteResponse"}}}},
          "400": {"$ref": "#/components/responses/Error"},
          "403": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/messages/trash": {
      "post": {
        "operationId": "trash_messages",
        "summary": "Move messages to Trash",
        "description": "Move one or more messages to the Trash mailbox (reversible; Proton purges Trash after 30 days). Use this to delete mail from INBOX, folders and labels; confirm with the user before calling it. Messages that are already in Trash (for example reached through another label) are skipped and reported in skipped_uids. For messages already in Spam or Trash, use delete_messages or empty_mailbox instead.",
        "tags": ["mail"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MessagesRef"}}}},
        "responses": {
          "200": {"description": "Trashed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusResponse"}}}},
          "403": {"$ref": "#/components/responses/Error"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer"}
    },
    "responses": {
      "Error": {
        "description": "Error",
        "content": {"application/json": {"schema": {"type": "object", "properties": {"error": {"type": "string"}}}}}
      }
    },
    "schemas": {
      "Address": {
        "type": "object",
        "properties": {
          "name": {"type": "string", "description": "Display name, if any"},
          "email": {"type": "string", "description": "Email address"}
        },
        "required": ["email"]
      },
      "Mailbox": {
        "type": "object",
        "properties": {
          "name": {"type": "string", "description": "Exact mailbox name to pass to other tools"},
          "kind": {"type": "string", "enum": ["system", "folder", "label"], "description": "system mailboxes (INBOX, Sent, ...), user folders ('Folders/...') or labels ('Labels/...')"},
          "role": {"type": "string", "description": "System role: inbox, sent, drafts, trash, archive, spam, starred, all; empty for custom folders and labels"},
          "total_messages": {"type": "integer"},
          "unread_messages": {"type": "integer"}
        },
        "required": ["name"]
      },
      "MessageSummary": {
        "type": "object",
        "properties": {
          "mailbox": {"type": "string"},
          "uid": {"type": "integer", "description": "Message id within the mailbox; pass with the mailbox to other tools"},
          "date": {"type": "string", "format": "date-time"},
          "from": {"type": "array", "items": {"$ref": "#/components/schemas/Address"}},
          "to": {"type": "array", "items": {"$ref": "#/components/schemas/Address"}},
          "subject": {"type": "string"},
          "message_id": {"type": "string", "description": "RFC 5322 Message-ID without angle brackets"},
          "read": {"type": "boolean"},
          "flagged": {"type": "boolean"},
          "answered": {"type": "boolean"},
          "has_attachments": {"type": "boolean"},
          "size_bytes": {"type": "integer"}
        }
      },
      "Message": {
        "allOf": [
          {"$ref": "#/components/schemas/MessageSummary"},
          {
            "type": "object",
            "properties": {
              "cc": {"type": "array", "items": {"$ref": "#/components/schemas/Address"}},
              "reply_to": {"type": "array", "items": {"$ref": "#/components/schemas/Address"}},
              "in_reply_to": {"type": "string"},
              "references": {"type": "array", "items": {"type": "string"}},
              "list_unsubscribe": {"type": "array", "items": {"type": "string"}, "description": "URLs (https or mailto) from the List-Unsubscribe header, if the sender provides one-click unsubscribe"},
              "body": {"type": "string", "description": "Plain-text body"},
              "body_truncated": {"type": "boolean", "description": "True when the body was cut to the configured maximum length"},
              "links": {"type": "array", "items": {"$ref": "#/components/schemas/Link"}, "description": "Hyperlinks from the HTML body with their anchor text (first 50)"},
              "attachments": {"type": "array", "items": {"$ref": "#/components/schemas/Attachment"}}
            }
          }
        ]
      },
      "Link": {
        "type": "object",
        "properties": {
          "text": {"type": "string", "description": "Anchor text"},
          "url": {"type": "string"}
        }
      },
      "Attachment": {
        "type": "object",
        "properties": {
          "filename": {"type": "string"},
          "content_type": {"type": "string"},
          "size_bytes": {"type": "integer"}
        }
      },
      "SearchRequest": {
        "type": "object",
        "properties": {
          "mailbox": {"type": "string", "description": "Mailbox to search. Defaults to INBOX.", "default": "INBOX"},
          "query": {"type": "string", "description": "Free-text search across headers and body"},
          "from": {"type": "string", "description": "Substring to match in the From header (name or address)"},
          "to": {"type": "string", "description": "Substring to match in the To header"},
          "subject": {"type": "string", "description": "Substring to match in the Subject header"},
          "since": {"type": "string", "description": "Only messages received on or after this date, YYYY-MM-DD"},
          "before": {"type": "string", "description": "Only messages received before this date, YYYY-MM-DD"},
          "unread_only": {"type": "boolean", "description": "Only return unread messages", "default": false},
          "limit": {"type": "integer", "description": "Maximum number of messages to return (newest first). Default 20.", "default": 20}
        }
      },
      "MessageRef": {
        "type": "object",
        "properties": {
          "mailbox": {"type": "string", "description": "Mailbox containing the message"},
          "uid": {"type": "integer", "description": "Message uid from search_messages"}
        },
        "required": ["mailbox", "uid"]
      },
      "MessagesRef": {
        "type": "object",
        "properties": {
          "mailbox": {"type": "string", "description": "Mailbox containing the messages"},
          "uids": {"type": "array", "items": {"type": "integer"}, "description": "One or more message uids from search_messages"}
        },
        "required": ["mailbox", "uids"]
      },
      "SendRequest": {
        "type": "object",
        "properties": {
          "to": {"type": "array", "items": {"type": "string"}, "description": "Recipient addresses, e.g. 'alice@example.com' or 'Alice <alice@example.com>'"},
          "cc": {"type": "array", "items": {"type": "string"}},
          "bcc": {"type": "array", "items": {"type": "string"}},
          "subject": {"type": "string"},
          "body": {"type": "string", "description": "Plain-text body"}
        },
        "required": ["to", "subject", "body"]
      },
      "ReplyRequest": {
        "type": "object",
        "properties": {
          "mailbox": {"type": "string", "description": "Mailbox containing the original message"},
          "uid": {"type": "integer", "description": "Uid of the original message"},
          "body": {"type": "string", "description": "Plain-text reply body (the original is not quoted automatically)"},
          "reply_all": {"type": "boolean", "description": "Reply to all original recipients instead of just the sender", "default": false}
        },
        "required": ["mailbox", "uid", "body"]
      },
      "FlagsRequest": {
        "type": "object",
        "properties": {
          "mailbox": {"type": "string"},
          "uids": {"type": "array", "items": {"type": "integer"}},
          "read": {"type": "boolean", "description": "true = mark read, false = mark unread, omit = unchanged"},
          "flagged": {"type": "boolean", "description": "true = star, false = unstar, omit = unchanged"}
        },
        "required": ["mailbox", "uids"]
      },
      "MoveRequest": {
        "type": "object",
        "properties": {
          "mailbox": {"type": "string", "description": "Mailbox the messages are currently in"},
          "uids": {"type": "array", "items": {"type": "integer"}},
          "destination": {"type": "string", "description": "Exact name of an existing mailbox to move into"}
        },
        "required": ["mailbox", "uids", "destination"]
      },
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "status": {"type": "string"},
          "mailbox": {"type": "string"},
          "deleted": {"type": "integer", "description": "Messages permanently deleted"}
        }
      },
      "LabelRequest": {
        "type": "object",
        "properties": {
          "mailbox": {"type": "string", "description": "Mailbox the messages are listed in"},
          "uids": {"type": "array", "items": {"type": "integer"}},
          "label": {"type": "string", "description": "Existing label name, e.g. 'Receipts' or 'Labels/Receipts'"}
        },
        "required": ["mailbox", "uids", "label"]
      },
      "LabelResponse": {
        "type": "object",
        "properties": {
          "status": {"type": "string"},
          "label": {"type": "string"},
          "labeled": {"type": "integer", "description": "Messages the label was added to"},
          "unlabeled": {"type": "integer", "description": "Messages the label was removed from"},
          "skipped_uids": {"type": "array", "items": {"type": "integer"}, "description": "Uids left untouched because they already had / did not have the label"}
        }
      },
      "SentResponse": {
        "type": "object",
        "properties": {
          "status": {"type": "string"},
          "from": {"$ref": "#/components/schemas/Address"},
          "to": {"type": "array", "items": {"$ref": "#/components/schemas/Address"}},
          "cc": {"type": "array", "items": {"$ref": "#/components/schemas/Address"}},
          "subject": {"type": "string"}
        }
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "status": {"type": "string"},
          "updated": {"type": "integer", "description": "Messages whose flags were changed"},
          "moved": {"type": "integer", "description": "Messages actually moved"},
          "skipped_uids": {"type": "array", "items": {"type": "integer"}, "description": "Uids left untouched because they were already in the destination"},
          "destination": {"type": "string"}
        }
      }
    }
  }
}